Fixes highlights causing value to be set instead of inner text
Problem occured when we highlight a option we add children to it and then
we count that as a nested list and can't be sure we can get the inner
text value and fallback to using the value of the option.

After we fix we check if the all the children are of type strong, and if
they are we set the inner text of the value.